parrot: Suppress avc denial for rkpdapp
Rkpdapp is an AOSP apk, which goes through android framework where
Perf hooks are present. These hooks are added in the framework to
identify key user activities like App launch, scroll, etc. Dontaudit
rule is added to supress the denail as there is no direct connection
between rkpdapp and perf hal.

Tests: The denial is gone with this change.
Change-Id: Ief98f15844211f01b892058c676dd285d1ba5df4